Navbar Test 2 Final
I wasn't able to find a fix for my issue. Because of this, I am opting to use my first method of making a navigation bar (despite it being a bit messy), just because I know how to use it and continuing to mess with this method when it is clearly not working would be a waste of my limited time. It's a little annoying, but this is the purpose of testing.