Grid Sizing Issue Fix
As it turns out, adding "1fr" to the end of the grid row template fixes it. I think this is because without it, the grid row template is automatically giving each row an equal size. However, by adding the 1fr, the auto is able to take the actual size of each row as the 1fr takes up the size of any remaining free space.