Image Issue
For some reason, the image absolutely refuses to be normal and centre itself to the centre of the div whilst having a height relative to the width of the source image. Instead, it chooses to stay stuck to the top of the div AND it chooses to inherit the height of the div rather than use its own height.