Image Issue Fix
As it turns out, I was using "justify-content" instead of "align-items" to try align the image to the centre. Not only that, but I also was doing something incredibly strange with the whole extra div within the div that added nothing, so I removed that. In the end, the image worked exactly as intended.