Complete Calendar Page Template
I successfully added a calendar to the calendar page, which is nice. The calendar page is probably gonna remain untouched from this point onwards though (other than basic styling to make it not look like a colourful mess), seeing as most of the styling would need to be done within the calendar spreadsheet itself - something in which the stakeholder has no intent of giving me access to.