Complete Contact Page Template
I believe I have now completed all the templates for each of the pages! Now, I need to get my stakeholder to check the website and see if they are happy with it and to see what they would like changed. After correcting any change the stakeholder may request, sprint 1 should be finished :)