Navbar Issue Fix + After Stakeholder Feedback
I figured that trying to do everything as a relative height just wasn't feasible with my level of knowledge, so I decided to just make the heights absolute. As I have the image for the logo, I know how large the width and height have to be, allowing this entire process to be easier. Once I had the actual height, I could just set the buttons to the same height as the logo and the flexbox would do all the rest of the work!