Members Page + Contact Page + Footer After Stakeholder Feedback
Seeing as each of the issues were very minor (the lack of padding on the sides and the footer having too many elements), I was able to quickly knock all these issues out the way. I believe now, I am able to show my stakeholder once again and see whether or not this is now ok!