Calendar Styling
I needed to style the pages, so I thought I'd get the simplest one done and out the way first. This one only really consisted of changing a couple of colours, but even then it looks really good compared to before! I did notice an issue with my original plan of colouring the webpages (where I planned to have it in main.css under grid-container), but luckily I didn't start relying on it before noticing the issue.