Footer Colour Change + Contact Entry Boxes Fix
My stakeholder asked for the footer to be darker, which I did very easily as the footer colour for every page is controlled by a singular css file. My stakeholder also asked for the entry boxes to have a little bit more width to them, which I also added. I also noticed a bug the form could be submitted without having filled in the required entry points, so I went ahead and fixed that too while I was still working on the contact page.