Moved Homepage Header
I moved the homepage header to within the hero banner. It took a little bit of work, but I successfully managed to make it look decently ok and resize with the page! In addition to that, I also had to change the way I made the hero banner, which might honestly be a blessing in disguise as I believe I can now change the size of the banner depending on the page size more easily, compared to the previous method where I wasn't too sure what I was going to do for a mobile page. Last thing I need to do now for the stakeholder is add an image to the main information, then I'll theoretically be done with sprint 1.