Homepage Additional Image
The stakeholder thought the information section was a bit boring, so I added an image. Surprisingly, it was far more of a headache than I anticipated. Fortunately though, I was able to sort it out fairly quickly, and now I have finished the home page once again. I also personally didn't like how the location details looked on the page seeing as I initially failed to make it look how I intended to design it, but here we are. The website is absolutely not suited for mobile just yet by any stretch of the imagination, but I'm currently more focused on what the website looks like on desktop.