Footer Revert Change + Minor Tweak
The stakeholder asked for the footer to get changed back, as he didn't like the look of the new footer - instead preferring the more seamless look. I also changed the width of the calendar ever so slightly as even though it wasn't bothering anyone else, it was bothering me greatly.