Copied W3schools Image Slideshow
I directly copied the code for the image slideshow, but I'm not entirely sure what it does. What I want to do now that I have this template code is change it a bit so that it fits what I want it to be, rather than leaving it the mess that it is.