Image Slideshow Further Testing
I created an entirely new page to see whether it was my code in particular that was causing the image to not load, or if it was the w3schools code which made it not load. As it turns out, it's the w3schools code that makes it not load properly, seeing as I directly copied the code and added nothing else to the new page.