Image Slideshow Fix Discovery?
I found that putting the javascript right down at the bottom rather than in the header causes it to work. I have a theory that the reason for this is because the javascript runs a certain line at the beginning that sets the image to the first one, but putting the javascript at the beginning means that the image is unable to be set as it hasn't even been put onto the page yet.