Image Slideshow Positioning
I successfully positioned the image slideshow in such a way where both the slideshow and dots at the bottom are lined up correctly and are both in the centre of their quadrant. Now, all that is really left is to put all the correct images in, clean it up a little, then I can show my stakeholder!