Navigation Bar Improvements
Two of the suggestions made were changing the colours of hovering over the navigation bar buttons (which admittedly looked kinda ugly before, I just never realised), and adding a transition to the colour. I didn't even know there was an option to give a transition to the colour until like just recently. I only spotted the transition attribute because I realised the arrow buttons on the image carousel had a transition too.