Contact Page Minor Fix
As it turns out, the page broke a little bit without me realising as when I added display: grid; to every page, it had the effect of causing a large gap between the info writing and the form on big screens. I believe I got lucky that both users have screens bigger than 1080p, as the bug doesn't show on screens 1080p or smaller (and I have been making this website on a laptop screen, making it really difficult to find the bug in the first place).