Semantic HTML
I recently discovered semantic HTML because one of my user testers was dyslexic and uses a screen reader on most websites, their screen reader not working all too well on my website. From googling, I found that screen readers tend to read the layout of the webpage straight from the HTML - relying on semantic HTML to truly understand how the page is actually set out. Because of this, I put in a bit of effort to try and learn a bit about semantic HTML so I could apply it to my own website. I'm not sure how well I applied the semantic HTML, but I hope it will at least be helpful to someone :)