USBWebserver Add + Test
One of the things needed to pass the level 3 DTS assessment is a database. I have never learned about databases before, as I always chose the programming topic when given the choice, so I have absolutely no clue what I'm doing. The website I'm using told me to install this program, which I believe I'm going to need for the database of the website. If it turns out I don't need this program, then I'll remove it.