Contact Page Basic Functionality
I've set up the contact page so that when filling in all the information on the page, it just gets sent to the database for admins to read. I decided to get this functionality out of the way as I realised it would probably be the much easier thing to do between a login button and this.