Basic Login Concept
I've watched a video and through that video, added a login system where all it really does is allows for people to log in and log out (although logging out requires visiting a certain page that isn't even on the navigation bar). This isn't how I want the login system to work in the end, but it is at least a start I can use to then build a more functional website.