Include Navigation Bar + Footer
After using the "include" function in PHP, I was curious as to whether or not it would work. To my surprise, it did flawlessly, meaning I can easily change the navigation bar and footer entirely from one page! What this means is that theoretically, it should be much easier to make some of the future changes that I want to make, as I only have to get them working for one file rather than every subsequent file.