Login Functionality
I've changed the login menu so that it functions better than what the one video did. All the login screen does is swap the login button to log in and log out depending on whether or not a user is logged in. I do wanna try add a little more functionality (such as accessing a members-only page), so that will likely be my next step. Either way, I'm amazed I have managed to achieve this much on the website, as I didn't think at the beginning of this project that I'd ever be able to make a login system in the first place.