Admin Page Test
I wanted to give some functionality to logging into an account, so I made it where it allows access to a hidden page for admins. This admin page doesn't do much, but it demonstrates a prototype that the stakeholder could use in the future for other features (such as a way for members to update their profiles, although that won't be something I do for this assessment). I still need to add some styling to the pages, so I'll quickly do that.