Login Popup Issue Fix
As it turns out, sleeping is very much a helpful thing when programming. I realised the reason why it wasn't working was because I failed to copy all the CSS across from W3Schools. Now that I actually have the code, I can now change it up a bit so that it better fits my specific requirements.