Files Cleanup
My file management was absolutely atrocious, with many files being entirely unnecessary and not even a part of the website in the first place. Because of that, I decided to go ahead and make the files a little bit more manageable. I've still yet to clean up my code though, so that will be my next step.