Mobile Friendlier Navigation Bar
I made the navigation bar slightly more mobile friendly. This time around, the login button takes up less room, so the navigation bar isn't quite so large on smaller screens. I also discovered that by creating an empty div with a style of "flex-grow: 1;" it allows for the login button to still remain on the opposite side of the navigation bar (without doing whatever I had going on before, I still don't understand why I did it like that...) Unfortunately, I'm still yet to figure out how to make the navigation turn into a hamburger menu when the screen width is small enough, but I don't believe it is too important of an issue for now, seeing as what I've got is easily an improvement.