Code Cleanup Sprint 3
Probably the heftiest code cleanup out of all of them. For this code cleanup, I went down a list and checked off to make sure I had everything. In the end, I did have to make some changes, but my code is otherwise ok. If all goes according to plan, the stakeholder should be ok with the state of sprint 3, and then I'll only need to make one more commit to update the readme once I can confirm everything is finished!