Update README.md
I updated the readme a little bit to make it clearer how to use just in case. It's probably not clearer at all, but it's good enough. Either way, this should be the very final commit for the assessment. From this point onwards, the code is in the hands of ECB Toastmasters.